Finished refactoring
