+++ src/api/collection_search.rs
@@ -1,11 +1,14 @@
-use super::*;
+use common::*;
+
+use api::Message;
+use svc;
 
 #[derive(PartialEq, Debug, Clone)]
 pub struct CollectionSearchRequest {
   pub node_id: NodeId,
 }
 
-impl api::Message for CollectionSearchRequest {
+impl Message for CollectionSearchRequest {
   type Protobuf = svc::CollectionSearchRequest;
 
   fn from_protobuf(protobuf: Self::Protobuf) -> Result<Self, api::Error> {
@@ -21,8 +24,8 @@ impl api::Message for CollectionSearchRequest {
   }
 
   #[cfg(test)]
-  fn required_fields() -> Self {
-    let node_id = NodeId::required_fields();
+  fn new_valid_test_instance() -> Self {
+    let node_id = NodeId::new_valid_test_instance();
     CollectionSearchRequest { node_id: node_id }
   }
 }
@@ -31,12 +34,12 @@ impl api::Message for CollectionSearchRequest {
 mod tests {
   use super::*;
 
-  use super::super::tests::*;
+  use api::tests::*;
 
   #[test]
-  fn collection_search_request_required_fields() {
+  fn collection_search_request_new_valid_test_instance() {
     test_required_fields::<CollectionSearchRequest, svc::CollectionSearchRequest>(&[|p| {
-      p.set_node_id(NodeId::required_fields().into_protobuf())
+      p.set_node_id(NodeId::new_valid_test_instance().into_protobuf())
     }]);
   }
 }
